Refactor RazorCodeActions
