+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/CodeActions/AddUsingsCodeActionHelper.cs
@@ -4,106 +4,46 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Razor.Language;
 using Microsoft.AspNetCore.Razor.Language.Components;
 using Microsoft.AspNetCore.Razor.Language.Extensions;
 using Microsoft.AspNetCore.Razor.Language.Legacy;
 using Microsoft.AspNetCore.Razor.Language.Syntax;
-using Microsoft.AspNetCore.Razor.LanguageServer.CodeActions.Models;
-using Microsoft.AspNetCore.Razor.LanguageServer.Common;
-using Microsoft.AspNetCore.Razor.LanguageServer.ProjectSystem;
-using Microsoft.CodeAnalysis.Razor;
-using Newtonsoft.Json.Linq;
 using OmniSharp.Extensions.LanguageServer.Protocol.Models;
 
 namespace Microsoft.AspNetCore.Razor.LanguageServer.CodeActions
 {
-    internal class AddUsingsCodeActionResolver : RazorCodeActionResolver
+    internal static class AddUsingsCodeActionHelper
     {
-        private readonly ForegroundDispatcher _foregroundDispatcher;
-        private readonly DocumentResolver _documentResolver;
-
-        public AddUsingsCodeActionResolver(ForegroundDispatcher foregroundDispatcher, DocumentResolver documentResolver)
+        internal static WorkspaceEdit CreateAddUsingWorkspaceEdit(string @namespace, RazorCodeDocument codeDocument, VersionedTextDocumentIdentifier codeDocumentIdentifier)
         {
-            _foregroundDispatcher = foregroundDispatcher ?? throw new ArgumentNullException(nameof(foregroundDispatcher));
-            _documentResolver = documentResolver ?? throw new ArgumentNullException(nameof(documentResolver));
-        }
-
-        public override string Action => LanguageServerConstants.CodeActions.AddUsing;
-
-        public override async Task<WorkspaceEdit> ResolveAsync(JObject data, CancellationToken cancellationToken)
-        {
-            if (data is null)
-            {
-                return null;
-            }
-
-            var actionParams = data.ToObject<AddUsingsCodeActionParams>();
-            if (actionParams is null)
-            {
-                return null;
-            }
-
-            var path = actionParams.Uri.GetAbsoluteOrUNCPath();
-
-            var document = await Task.Factory.StartNew(() =>
-            {
-                _documentResolver.TryResolveDocument(path, out var documentSnapshot);
-                return documentSnapshot;
-            }, cancellationToken, TaskCreationOptions.None, _foregroundDispatcher.ForegroundScheduler).ConfigureAwait(false);
-            if (document is null)
-            {
-                return null;
-            }
-
-            var text = await document.GetTextAsync().ConfigureAwait(false);
-            if (text is null)
-            {
-                return null;
-            }
-
-            var codeDocument = await document.GetGeneratedOutputAsync().ConfigureAwait(false);
-            if (codeDocument.IsUnsupported())
-            {
-                return null;
-            }
-
-            if (!FileKinds.IsComponent(codeDocument.GetFileKind()))
-            {
-                return null;
-            }
-
-            var codeDocumentIdentifier = new VersionedTextDocumentIdentifier() { Uri = actionParams.Uri };
-            var documentChanges = new List<WorkspaceEditDocumentChange>();
-
             /* The heuristic is as follows:
-             * 
-             * - If no @using, @namespace, or @page directives are present, insert the statements at the top of the 
-             *   file in alphabetical order.
-             * - If a @namespace or @page are present, the statements are inserted after the last line-wise in
-             *   alphabetical order.
-             * - If @using directives are present and alphabetized with System directives at the top, the statements
-             *   will be placed in the correct locations according to that ordering.
-             * - Otherwise it's kinda undefined; it's only geared to insert based on alphabetization.
-             * 
-             * This is generally sufficient for our current situation (inserting a single @using statement to include a
-             * component), however it has holes if we eventually use it for other purposes. If we want to deal with 
-             * that now I can come up with a more sophisticated heuristic (something along the lines of checking if 
-             * there's already an ordering, etc.).
-             */
+            *
+            * - If no @using, @namespace, or @page directives are present, insert the statements at the top of the
+            *   file in alphabetical order.
+            * - If a @namespace or @page are present, the statements are inserted after the last line-wise in
+            *   alphabetical order.
+            * - If @using directives are present and alphabetized with System directives at the top, the statements
+            *   will be placed in the correct locations according to that ordering.
+            * - Otherwise it's kinda undefined; it's only geared to insert based on alphabetization.
+            *
+            * This is generally sufficient for our current situation (inserting a single @using statement to include a
+            * component), however it has holes if we eventually use it for other purposes. If we want to deal with
+            * that now I can come up with a more sophisticated heuristic (something along the lines of checking if
+            * there's already an ordering, etc.).
+            */
+            var documentChanges = new List<WorkspaceEditDocumentChange>();
             var usingDirectives = FindUsingDirectives(codeDocument);
             if (usingDirectives.Count > 0)
             {
                 // Interpolate based on existing @using statements
-                var edits = GenerateSingleUsingEditsInterpolated(codeDocument, codeDocumentIdentifier, actionParams.Namespace, usingDirectives);
+                var edits = GenerateSingleUsingEditsInterpolated(codeDocument, codeDocumentIdentifier, @namespace, usingDirectives);
                 documentChanges.Add(edits);
             }
             else
             {
                 // Just throw them at the top
-                var edits = GenerateSingleUsingEditsAtTop(codeDocument, codeDocumentIdentifier, actionParams.Namespace);
+                var edits = GenerateSingleUsingEditsAtTop(codeDocument, codeDocumentIdentifier, @namespace);
                 documentChanges.Add(edits);
             }
