+++ src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/CodeActions/ExtractToCodeBehindCodeActionProviderTest.cs
@@ -15,6 +15,7 @@ using Moq;
 using OmniSharp.Extensions.LanguageServer.Protocol.Models;
 using Xunit;
 using Range = OmniSharp.Extensions.LanguageServer.Protocol.Models.Range;
+using Microsoft.CodeAnalysis.Text;
 
 namespace Microsoft.AspNetCore.Razor.LanguageServer.Test.CodeActions
 {
@@ -243,7 +244,13 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Test.CodeActions
                 document.GetGeneratedOutputAsync() == Task.FromResult(codeDocument) &&
                 document.GetTextAsync() == Task.FromResult(codeDocument.GetSourceText()));
 
-            return new RazorCodeActionContext(request, documentSnapshot, codeDocument, location, supportsFileCreation);
+
+            var sourceText = Mock.Of<SourceText>();
+
+            var context = new RazorCodeActionContext(request, documentSnapshot, codeDocument, sourceText, supportsFileCreation);
+            context.Location = location;
+
+            return context;
         }
     }
 }
